+++ src/classy_blocks/construct/operations/extrude.py
@@ -1,3 +1,5 @@
+from typing import Union
+
 import numpy as np
 
 from classy_blocks.construct.flat.face import Face
@@ -7,12 +9,18 @@ from classy_blocks.types import VectorType
 
 
 class Extrude(Loft):
-    """Takes a Face and extrudes it in given extrude_direction"""
+    """Takes a Face and extrudes it by 'amount'.
+    If 'amount' is float, the extrude direction is normal to 'base'.
+    """
 
-    def __init__(self, base: Face, extrude_vector: VectorType):
+    def __init__(self, base: Face, amount: Union[float, VectorType]):
         self.base = base
-        self.extrude_vector = np.asarray(extrude_vector)
 
-        top_face = base.copy().translate(self.extrude_vector)
+        if isinstance(amount, float):
+            extrude_vector = self.base.normal * amount
+        else:
+            extrude_vector = np.asarray(amount)
+
+        top_face = base.copy().translate(extrude_vector)
 
         super().__init__(base, top_face)