+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1. 1. 0] Default Extrude Direction
+### Added
+- Extrude now takes a vector of a float. If a float is given, direction is normal of the base face.
+
 ## [1. 0. 0] Refactor
 A complete overhaul of all objects in an attempt to create a proper SOLID-obeying
 package with type hinting, static typing and no python-ish duck-typing hacks.